+++ web/components/item/index.ts
@@ -1,2 +1,4 @@
 export * from "./ItemCard"
+export * from "./ItemScanHistory"
+export * from "./ItemScanHistoryItem"
 export * from "./types"
